refactor: move variables to constants dir
