+++ src/app/Home.tsx
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import { chatbotConfig } from "@src/lib/bot/chatbot-config";
+import { chatbotConfig } from "@constants/bot/chatbot-config";
 
 // components
 import Button from "@components/ui/Button";
