+++ src/providers/ChatbotProvider.tsx
@@ -13,7 +13,7 @@ import {
 } from "@src/shared/ts/type";
 
 // lib
-import { chatbotConfig } from "@src/lib/bot/chatbot-config";
+import { chatbotConfig } from "@constants/bot/chatbot-config";
 
 const botProfileCollectionRef = collection(db, "profile");
